fix(nlu-server): fix send training_time using db config
