+++ packages/nlu-server/src/infrastructure/training-repo/in-memory-training-repo.ts
@@ -6,7 +6,6 @@ import Bluebird from 'bluebird'
 import _ from 'lodash'
 import moment from 'moment'
 import ms from 'ms'
-import { BaseWritableTrainingRepo, BaseTrainingRepository } from './base-training-repo'
 
 import {
   Training,
@@ -14,7 +13,8 @@ import {
   TrainingState,
   TrainingRepository,
   TrainingTrx,
-  WrittableTrainingRepository
+  WrittableTrainingRepository,
+  TrainingListener
 } from './typings'
 
 const KEY_JOIN_CHAR = '\u2581'
@@ -26,16 +26,23 @@ type TrainEntry = TrainingState & { updatedOn: Date } & {
   dataset: TrainInput
 }
 
-class WrittableTrainingRepo extends BaseWritableTrainingRepo implements WrittableTrainingRepository {
+class InMemoryWrittableTrainingRepo implements WrittableTrainingRepository {
+  private _listeners: TrainingListener[] = []
   private _trainSessions: {
     [key: string]: TrainEntry
   } = {}
 
-  constructor(logger: Logger) {
-    super(logger.sub('training-repo'))
+  private _janitorIntervalId: NodeJS.Timeout | undefined
+
+  constructor(private _logger: Logger) {}
+
+  public addListener(listener: TrainingListener) {
+    this._listeners.push(listener)
   }
 
-  private _janitorIntervalId: NodeJS.Timeout | undefined
+  public removeListener(listenerToRemove: TrainingListener) {
+    _.remove(this._listeners, (listener) => listener === listenerToRemove)
+  }
 
   public async initialize(): Promise<void> {
     this._janitorIntervalId = setInterval(this._janitor.bind(this), JANITOR_MS_INTERVAL)
@@ -65,7 +72,7 @@ class WrittableTrainingRepo extends BaseWritableTrainingRepo implements Writtabl
   }
 
   public async set(training: Training): Promise<void> {
-    await super.set(training)
+    this._onTrainingEvent(training)
     const key = this._makeTrainingKey(training)
     this._trainSessions[key] = { ...training, updatedOn: new Date() }
   }
@@ -119,15 +126,33 @@ class WrittableTrainingRepo extends BaseWritableTrainingRepo implements Writtabl
     const modelId = NLUEngine.modelIdService.fromString(stringId)
     return { modelId, appId }
   }
+
+  private _onTrainingEvent(training: Training) {
+    this._listeners.forEach((listener) => {
+      // The await keyword isn't used to prevent a listener from blocking the training repo
+      listener(training).catch((e) =>
+        this._logger.attachError(e).error('an error occured in the training repository listener')
+      )
+    })
+  }
 }
 
-export class InMemoryTrainingRepo extends BaseTrainingRepository<WrittableTrainingRepo> implements TrainingRepository {
+export class InMemoryTrainingRepo implements TrainingRepository {
   private _trxQueue: LockedTransactionQueue<void>
+  private _writtableTrainingRepository: InMemoryWrittableTrainingRepo
 
   constructor(logger: Logger) {
-    super(logger, new WrittableTrainingRepo(logger))
     const logCb = (msg: string) => logger.sub('trx-queue').debug(msg)
     this._trxQueue = makeInMemoryTrxQueue(logCb)
+    this._writtableTrainingRepository = new InMemoryWrittableTrainingRepo(logger)
+  }
+
+  public addListener(listener: TrainingListener) {
+    this._writtableTrainingRepository.addListener(listener)
+  }
+
+  public removeListener(listener: TrainingListener) {
+    this._writtableTrainingRepository.removeListener(listener)
   }
 
   public async initialize(): Promise<void> {
